RODI-54: old or repeat run should be first logged

+++ api/src/main/java/org/openmrs/module/systemmonitor/api/impl/SystemMonitorServiceImpl.java
@@ -731,11 +731,11 @@ public class SystemMonitorServiceImpl extends BaseOpenmrsService implements Syst
 
 	@Override
 	public JSONObject pushMonitoredDataToDHIS() {
-		updatePreviouslySubmittedSMTData();
-		return runSMTEvaluatorAndLogOrPushData();
+		updatePreviouslySubmittedOrMissedSMTData();
+		return runSMTEvaluatorAndLogOrPushData(false);
 	}
 
-	private JSONObject runSMTEvaluatorAndLogOrPushData() {
+	private JSONObject runSMTEvaluatorAndLogOrPushData(Boolean passedOrRepeatRun) {
 		JSONObject response = null;
 		JSONObject dataToBePushed = getDataToPushToDHIS();
 
@@ -755,15 +755,19 @@ public class SystemMonitorServiceImpl extends BaseOpenmrsService implements Syst
 				&& dataToBePushed.getJSONArray("dataValues") != null
 				&& dataToBePushed.getJSONArray("dataValues").length() > 0) {
 			try {
-				response = CurlEmulator.post(dhisPostUrl, dataToBePushed, dhisUserName, dhisPassword);
-				if (response != null) {
-					logCurlEmulatorPostResponse(response.toString());
-					if (backupFile.exists()) {
-						backupFile.delete();
-					}
-					pushPreviouslyFailedDataWhenOutOfInternet();
-				} else {
+				if(passedOrRepeatRun) {
 					backupSystemMonitorDataToPush(dataToBePushed);
+				} else {
+					response = CurlEmulator.post(dhisPostUrl, dataToBePushed, dhisUserName, dhisPassword);
+					if (response != null) {
+						logCurlEmulatorPostResponse(response.toString());
+						if (backupFile.exists()) {
+							backupFile.delete();
+						}
+						pushPreviouslyFailedDataWhenOutOfInternet();
+					} else {
+						backupSystemMonitorDataToPush(dataToBePushed);
+					}
 				}
 
 			} catch (UnknownHostException e) {
@@ -1373,29 +1377,31 @@ public class SystemMonitorServiceImpl extends BaseOpenmrsService implements Syst
 	 * this functionality can only run until 31st/march/2016 which means by this
 	 * date all Sites must have been upgraded
 	 */
-	private void updatePreviouslySubmittedSMTData() {
+	private void updatePreviouslySubmittedOrMissedSMTData() {
 		try {
 			SimpleDateFormat sdf = dao.getSdf();
-			Date supportedUntil = sdf.parse("2017-03-31 00:00:00");
+			String s = Context.getAdministrationService().getGlobalProperty(SystemMonitorConstants.SMT_EVAL_SD_SUPPORT_TO);
+			Date supportedUntil = StringUtils.isNotBlank(s) ? sdf.parse(s) : null;
 			Calendar date = Calendar.getInstance(Context.getLocale());
 			Calendar today = Calendar.getInstance(Context.getLocale());
 			GlobalProperty evalDateGp = Context.getAdministrationService()
 					.getGlobalPropertyObject(ConfigurableGlobalProperties.EVALUATION_AND_REPORTING_DATE);
 			File smtBackUpDirectory = SystemMonitorConstants.SYSTEMMONITOR_DIRECTORY;
 			File smtBackUpDataDirectory = SystemMonitorConstants.SYSTEMMONITOR_BACKUPFOLDER;
-			String dateStr = evalDateGp.getPropertyValue();
+			String dateStr = evalDateGp != null ? evalDateGp.getPropertyValue() : null;
 
-			if (evalDateGp != null && StringUtils.isNotBlank(dateStr) && smtBackUpDirectory.exists()
-					&& (smtBackUpDirectory.isDirectory() && smtBackUpDirectory.listFiles().length > 0)
-					&& ((smtBackUpDataDirectory.exists() && smtBackUpDataDirectory.isDirectory()))
-					&& (smtBackUpDataDirectory.listFiles().length > 0)
-					&& supportedUntil.after(today.getTime())) {
-
+			if(smtBackUpDirectory == null || !smtBackUpDirectory.exists())
+				smtBackUpDirectory.mkdir();
+			if(smtBackUpDataDirectory == null || !smtBackUpDataDirectory.exists())
+				smtBackUpDataDirectory.mkdirs();
+				
+			if (supportedUntil != null && StringUtils.isNotBlank(dateStr) && supportedUntil.before(today.getTime())) {
 				date.setTime(sdf.parse(dateStr));
 				while (today.after(date)) {
 					// eliminate weekend days
 					if (date.get(Calendar.DAY_OF_WEEK) != 1 && date.get(Calendar.DAY_OF_WEEK) != 7) {
-						runSMTEvaluatorAndLogOrPushData();
+						System.out.println("\nINFO - Running SMT for Date: " + dateStr);
+						runSMTEvaluatorAndLogOrPushData(true);
 					}
 					date.add(Calendar.DAY_OF_YEAR, 1);
 					dateStr = sdf.format(date.getTime());
@@ -1405,11 +1411,14 @@ public class SystemMonitorServiceImpl extends BaseOpenmrsService implements Syst
 				// finally
 				evalDateGp.setPropertyValue("");
 				Context.getAdministrationService().saveGlobalProperty(evalDateGp);
+			} else {
+				System.out.println("FAILURE:::::::" + supportedUntil + ":::::" + dateStr);
 			}
 		} catch (ParseException e) {
 			e.printStackTrace();
 		}
 	}
+
 
 	@Override
 	public Integer rwandaGetTotalActivePatients_AtleastEightMonthsARVTreatment() {
